+++ client/Visit.py
@@ -63,59 +63,61 @@ class Visit(Menu):
         self.show_res_win("Get a visit")
         visit = self.get_input(7, 4, "Give visit name")
 
-        resp = requests.get(self._url + "visits/" + visit + '/')
-
         try:
+            resp = requests.get(self._url + "visits/" + visit + '/')
             body = resp.json()
+            string = self.parse_visit(body)
+            lc = string.count('\n')
+            msg = "Visit: "
+            self.show_res(string, lc, msg)
         except ValueError:
             self.show_res("No visit found!", 1, "Oh noes")
             return
-
-        string = self.parse_visit(body)
-        lc = string.count('\n')
-
-        msg = "Visit: "
-        self.show_res(string, lc, msg)
+        except KeyError:
+            return
 
     def post_visit(self):
-        self.show_res_win("Post a visit")
-        visit = self.get_input(7, 4, "Give visit name")
-        mokki = self.get_input(10, 4, "Give mokki name")
-        start = self.get_input(13, 4,
-                               "Give the starting time of visit " +
-                               "(YYYY-MM-DDThh:mm:ss+00:00)")
-        end = self.get_input(16, 4,
-                             "Give the ending time of visit " +
-                             "(YYYY-MM-DDThh:mm:ss+00:00)")
-        part_str = self.get_input(19, 4,
-                                  "Give participants, separated by ;")
-        ps = part_str.split(';')
-
-        # The date is PITA to insert manually so,
-        # append the hour value if only date is inserted :)
         try:
-            left, right = start.split('T')
-        except ValueError:
-            start += "T12:00:00+00:00"
+            self.show_res_win("Post a visit")
+            visit = self.get_input(7, 4, "Give visit name")
+            mokki = self.get_input(10, 4, "Give mokki name")
+            start = self.get_input(13, 4,
+                                   "Give the starting time of visit " +
+                                   "(YYYY-MM-DDThh:mm:ss+00:00)")
+            end = self.get_input(16, 4,
+                                 "Give the ending time of visit " +
+                                 "(YYYY-MM-DDThh:mm:ss+00:00)")
+            part_str = self.get_input(19, 4,
+                                      "Give participants, separated by ;")
+            if visit == '' or mokki == '':
+                return
 
-        try:
-            left, right = end.split('T')
-        except ValueError:
-            end += "T13:00:00+00:00"
+            # The date is PITA to insert manually so,
+            # append the hour value if only date is inserted :)
+            try:
+                ps = part_str.split(';')
+                left, right = start.split('T')
+            except ValueError:
+                start += "T12:00:00+00:00"
 
-        data = {
-                "visit_name": visit,
-                "mokki_name": mokki,
-                "time_start": start,
-                "time_end": end,
-                "participants": ps
-        }
+            try:
+                left, right = end.split('T')
+            except ValueError:
+                end += "T13:00:00+00:00"
 
-        try:
+            data = {
+                    "visit_name": visit,
+                    "mokki_name": mokki,
+                    "time_start": start,
+                    "time_end": end,
+                    "participants": ps
+            }
+
             r = requests.post(self._url + "visits/",
                               data=json.dumps(data),
                               headers={"Content-type": "application/json"})
-
+        except KeyError:
+            return
         except Exception as e:
             e = str(e)
             lc = e.count("\n")
@@ -123,44 +125,48 @@ class Visit(Menu):
             return
 
         if r.status_code != 201:
-            print(r)
-            # msg = str(r.json()['@error']['@message'])
-            # err_code = str(r.status_code)
-            # lc = msg.count('\n')
-            # self.show_res(msg, lc, err_code)
+            msg = str(r.json()['@error']['@message'])
+            err_code = str(r.status_code)
+            lc = msg.count('\n')
+            self.show_res(msg, lc, err_code)
 
     def edit_visit(self):
-        self.show_res_win("Edit a visit")
-        visit = self.get_input(7, 4, "Enter the visit to modify")
-        visit_new = self.get_input(10, 4, "Enter new name to visit")
-        mokki = self.get_input(13, 4, "Enter new name to mokki")
-        start = self.get_input(16, 4, "New start time " +
-                               "(YYYY-MM-DDThh:mm:ss+00:00")
-        end = self.get_input(19, 4, "New end time " +
-                             "(YYYY-MM-DDThh:mm:ss+00:00")
-
-        # Append the hour value if only date is inserted :)
         try:
-            left, right = start.split('T')
-        except ValueError:
-            start += "T12:00:00+00:00"
+            self.show_res_win("Edit a visit")
+            visit = self.get_input(7, 4, "Enter the visit to modify")
+            visit_new = self.get_input(10, 4, "Enter new name to visit")
+            mokki = self.get_input(13, 4, "Enter new name to mokki")
+            start = self.get_input(16, 4, "New start time " +
+                                   "(YYYY-MM-DDThh:mm:ss+00:00")
+            end = self.get_input(19, 4, "New end time " +
+                                 "(YYYY-MM-DDThh:mm:ss+00:00")
 
-        try:
-            left, right = end.split('T')
-        except ValueError:
-            end += "T13:00:00+00:00"
+            if visit == '' or mokki == '' or visit_new == '':
+                return
 
-        data = {
-                "visit_name": visit_new,
-                "mokki_name": mokki,
-                "time_start": start,
-                "time_end": end
-        }
+            # Append the hour value if only date is inserted :)
+            try:
+                left, right = start.split('T')
+            except ValueError:
+                start += "T12:00:00+00:00"
 
-        try:
+            try:
+                left, right = end.split('T')
+            except ValueError:
+                end += "T13:00:00+00:00"
+
+            data = {
+                    "visit_name": visit_new,
+                    "mokki_name": mokki,
+                    "time_start": start,
+                    "time_end": end
+            }
+
             r = requests.put(self._url + "visits/{}/".format(visit),
                              data=json.dumps(data),
                              headers={"Content-type": "application/json"})
+        except KeyError:
+            return
 
         except Exception as e:
             e = str(e)
@@ -184,6 +190,8 @@ class Visit(Menu):
     def delete_visit(self):
         self.show_res_win("Delete a visit")
         visit = self.get_input(7, 4, "Enter a visit to be deleted")
+        if visit == '':
+            return
         try:
             r = requests.delete(self._url + "visits/{}/".format(visit))
         except Exception as e:
@@ -193,7 +201,10 @@ class Visit(Menu):
             return
 
         if r.status_code != 204:
-            msg = str(r.json()['@error']['@message'])
+            try:
+                msg = str(r.json()['@error']['@message'])
+            except Exception:
+                msg = str(r.text)
             err_code = str(r.status_code)
             lc = msg.count('\n')
             self.show_res(msg, lc, err_code)
